nios2: redirect exception vector outside sdram
The default exception vector was located at sdram. But some designs
locate the exception vector at onchip memory or sram, and the
default exception hook won't work without regenerate hardware.

This patch copy a sequence of instructions to the hardwired exception
location. So that exceptions are redirected to inthandler at sdram.

FIXME: It won't work if the exception slave is not writable. An error
should be generated then.

Signed-off-by: Thomas Chou <thomas@wytron.com.tw>