nios2: flush cache in ptrace poke text
The nios2 arch has seperated icache and dcache. We have to flush cache
to change text memory.

It is not necessary to flush usr and regs, as they will be accessed by
data master of nios2 exclusively. Only changes in text/code should be
flushed, as the code will be passed from data master to memory, and then
to instruction master.

Signed-off-by: Thomas Chou <thomas@wytron.com.tw>