nios2: enable 'make AFLAGS=...' to add additional options to AS
Follow the patch from Sam Ravnborg <sam@ravnborg.org>,

The variable AFLAGS is a wellknown variable and the usage by
kbuild may result in unexpected behaviour.
On top of that several people over time has asked for a way to
pass in additional flags to gcc.

This patch replace use of AFLAGS with KBUILD_AFLAGS all over
the tree.

Signed-off-by: Thomas Chou <thomas@wytron.com.tw>