nios2: system timer input clock frequency
frd66 identified a bug that the frequency declared in nios2_system.h was
wrong, which was the external one that go to the input of in-system PLL.
He submitted a patch to use internal cpu clock.

But the system timer input does not always come from cpu clock. I
refined his patch and look for the real clock input to the system timer.

na_cpu_clock_freq is added to display correct /proc/cpuinfo.

Signed-off-by: Thomas Chou <thomas@wytron.com.tw>