nios2: memcpy with stream-lined memory access
As sdram has long access latency, the total access time can be reduced
with sequential access in the same page.

This patch unrolls the copy loop and stream-lines the memory access as
(read x8 + write x8) for aligned source and destination, or
(read x4 + write x4) for unalinged, using registers for buffer.

Signed-off-by: Thomas Chou <thomas@wytron.com.tw>