nios2: flush data cache with address matching
Change the instruction to flush data cache to "flushda".

NiosII instruction spec: "flushda" flushes the cache line only if the
addressed memory location is currently cached. By contrast, the "flushd"
instruction flushes the cache line even if the addressed memory location
is not cached.

The flush range for cache_invalidate_lines() is limited to the cache
size, so that the time spent can be reduced.

Signed-off-by: Thomas Chou <thomas@wytron.com.tw>