nios2: resolve section overlap linkage error
Update the link script, based on the Blackfin's, to resolve section
overlap error.

The PAGE_SIZE and THREAD_SIZE defs in page.h are used instead of
numeric constant.

The code moving loop in head.S is removed, because a boot loader is
always used.

Note that the bss section is now before the data section. We want the
init sections to be freed without fragmentation.

Signed-off-by: Thomas Chou <thomas@wytron.com.tw>