nios2: Applied MTD CFI-driver patch from Mictrotronix 2.6.11 port
This patch is required to get flash erase (and thus writing) to work properly with the Avalon bus.
The driver relies on monitoring the 'toggling-bit' (toggles between subsequent reads)
of the data-bus to determine if an erase command has completed. The Avalon memory-mapped
interface will always read 32bit from the slave independent of the requested read-size.
A read will then generate 2 read-cycles on a 16bit flash-interface, and 4 cycles on an 8bit
interface - which makes detection of the 'toggling-bit' impossible. The patch instead uses
the 'polling-bit' along with the read data to determine if the erase cycle has completed.

The Avalon documentation states that the number of read-cycles should be dependent on the
masters byte-enables so the observed behavior is clearly a bug.
Eventually this bug should be fixed in HW, and this patch obsoleted.

Signed-off-by: Atle Nissestad <atle@nissestad.no>