Makefile: support external kernel source
From: Thomas Chou <thomas@wytron.com.tw>

Add support of external kernel source and build kernel outside the
source tree using the "O=" option.

Sometime we may want to build kernel from an external source such as a
git repository outside the uclinux-dist directory. Though we could use
git submodule of kernel inside a uclinux-dist git checkout, it is rather
limited in functionality. It will be more useful and cleaner to do it
outside.

We add a new variable LINUXSRC, which specifies the relative path of the
external kernel source from uClinux-dist ROOTDIR. This variable can be
defined in vendor config.arch or specified in command line. If it is not
defined, it will default to LINUXDIR and the rest does not change.

When LINUXSRC is set and different from LINUXDIR, the kernel build will
be output to LINUXDIR using the "O=" option. The .config and kernel
objects will go to the LINUXDIR during the build as before.

For example,
LINUXDIR is linux-2.6.x for 2.6 kernel.
And we put linux-kernel git dir at the same level as uclinux-dist dir,
LINUXSRC  = ../linux-kernel (defined in config.arch or command line)

Other usage of "$(MAKEARCH_KERNEL) -C " were changed to $(LINUXSRC),
such as vendor.mak.

For a uclinux-dit git checkout, there is no linux-2.6.x dir. So one will
be created, if it doesn't exist.

Signed-off-by: Thomas Chou <thomas@wytron.com.tw>


git-svn-id: svn://localhost/svn/uclinux-dist/trunk@8286 3e2a5524-39e8-4cb9-84a9-c49da69cba3a