cross-pkg-config: switch to PKG_CONFIG_SYSROOT_DIR (kind of)
The latest pkg-config supports PKG_CONFIG_SYSROOT_DIR which handles -L/-I
munging on the fly, but it's slightly broken.  So put all the pieces in
place for the next upstream release and in the mean time, do the munging
ourselves on the fly via sed.

This way packages that want runtime paths compiled in don't get the build
paths that were munged in the .pc file.


git-svn-id: svn://localhost/svn/uclinux-dist/trunk@9366 3e2a5524-39e8-4cb9-84a9-c49da69cba3a