[#6210] bfin_preempt_latency: add a simple driver for preempt latency test
bsong have added preempt to our kernel in commit 9079, add a simple driver
in order to test it.

This driver use TIMER2 to generate interrupt every seconds,the timer isr
handle will record the time and send SIGIO to userapp preempt_test which will
read the recorded time when scheduled and compare with the actual time to
calculate the latency.



git-svn-id: svn://localhost/svn/uclinux-dist/trunk@9820 3e2a5524-39e8-4cb9-84a9-c49da69cba3a