ftpd: add daemon (-D) nommu support
The current daemon design of ftpd is to:
    - loop in server_mode() waiting for a connection
    - fork a child and return to main() to process like inetd
    - have parent continue looping in server_mode()

On a nommu system where we can only vfork(), the child returning
from server_mode() corrupts the stack and messes up the the parent.
So rather than having the child return, exec a new ftpd process in
the normal inetd mode.  This also fixes the problem in nommu where
a vfork-ed child pauses the parent until it either exits or execs
something.


git-svn-id: svn://localhost/svn/uclinux-dist/trunk@9964 3e2a5524-39e8-4cb9-84a9-c49da69cba3a