at: change fork() to vfork() for atd
this isnt entirely correct as there are many points here which can call exit(), but they are all nested fairly deeply and this code will take quite a bit of rewrite to make it nommu clean.  since this app sucks and isnt commonly used, dont bother for now in doing so.

git-svn-id: svn://localhost/svn/uclinux-dist/trunk@10107 3e2a5524-39e8-4cb9-84a9-c49da69cba3a